requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
boto3==1.34.34
python-dotenv==1.0.1
openai>=1.14.0
python-multipart==0.0.9
requests==2.31.0 